requirements.txt
requests==2.22.0
pymssql==2.1.4
pytz==2019.3
paramiko==2.6.0
suds_jurko==0.6
